Add support for slds-max-size-show
Fix #263 by adding support for slds-max-size-show.

Breakpoints are set so that visibility occurs at one pixel below the
min-breakpoint. This keeps the current definition of the “max”
attribute being exclusive. This means an element will be shown for
widths up to but not including the pixel value of the max-size
specified.